refactor: merge Client addon to HZClient
