+++ src/djsAddon/index.ts
@@ -19,7 +19,6 @@
  */
 
 export * from './BaseChannel';
-export * from './Client';
 export * from './EmbedBuilder';
 export * from './GuildMember';
 export * from './User';
